Make Compute Package content sensitve
When providing the content directly to a fastly compute service

```
resource "fastly_service_compute" "my_service" {
  // ...
  package {
    content = data.some_content_provider.content
  }
}
```

It will output the full base64 content in the terminal and other
interfaces for terraform (e.g. atlantis) when planning the change.

Because packages typically range in megabytes, that's a lot of
(useless) information to have displayed.

While "sensitive" might not be exactly true (it's not a password), the
[documentation][schema-doc] for sensitive in the Schema states

> It should be used for password or other values which should be hidden.

This would fall in the category "other values which should be hidden".

It can be worked around locally by passing the contents to the
[sensitive][sensitive] function

```
package {
  content = sensitive(data.some_content_provider.content)
}
```

[schema-doc]: https://pkg.go.dev/github.com/hashicorp/terraform-plugin-sdk/v2@v2.31.0/helper/schema#Schema.Sensitive
[sensitive]: https://developer.hashicorp.com/terraform/language/functions/sensitive