feat(image_optimizer_default_settings): add Image Optimizer default settings API (#832)
* Temporarily vendor go-fastly from Image Optimizer default settings API PR

* Add Image Optimizer default settings API

* Generate docs.

* Handle case where reading image optimizer settings on a service without IO enabled.

* Fix "invalid key for element" error caused by missing "name" key

* PR cleanup suggestion.

* Include default values.

I was initially trying to get Terraform to ignore settings that wern't
set in the configuration, leaving them at their previously configured
value. This seems to be counter to Terraform's design, though, so I've
relented & included the current default values for each setting here.

It seems like maybe we could have done that for the string values, and
non-0 integers, but it seems like a nonstarter for booleans
(https://github.com/hashicorp/terraform-plugin-sdk/issues/817), and
I didn't want to go against Terraform's design too much.

* Update go-fastly vendored from PR

* Fix lints + add debug message.

* One more vendor update.

* Update jpeg_type, and with that, finally have a passing test!

* Update vendored code again.

* Update field documentation.

Part of this is updating field with descriptions from the final
api-documentation PR. The other part is just my messing with stuff to
try to make it look better.

* Update generated docs

(also, move image_optimizer_default_settings guide from docs to templates)

* Change Delete so it resets all settings to default.

* Overhaul Image Optimizer default settings test.

* Update vendored go-fastly to 9.4.0.

* Add debug message for ignoring image-optimizer-disabled error on deletion.

This is working well! Just wanted to add this to double-confirm it's
doing what I expect.

* Remove toolchain directive.

* Undo go version bump.

Not sure why updating had caused this, but this should still work.