feat: support mutual_authentication_id in domain activation at creation (#875)
* chore: add test case for activation with mTLS

This adds a (failing) test-case for creating a domain activation
with a mutual authentication configuration. Currently the test
case fails because the check on the resource shows it does
not match the state.

* fix: Update domain activation for mTLS after creation

The mutual_authentication_id can only be set once an activation has
already been created. This commit implements the Update call
immediately after the Create call if the mutual_authentication_id
is set in the resource configuration.

It fixes the failing test case added in the previous commit.

---------

Co-authored-by: Kevin P. Fleming <kpfleming@users.noreply.github.com>