breaking(product_enablement): Remove support for NGWAF product. (#893)
The support added for the NGWAF product in version 5.14.0 did not
work, and is being removed so that it can be redesigned. Since the
feature was non-operational, no users of the provider could have
relied on it, so the major version number is not being incremented in
spite of the breaking nature of the change.