Remove 'vendor' directory and all processes related to it.
With modern version of Go, vendoring dependencies only provides value
if there is a need to be able to build off-line, or if one or more
dependencies have disappeared (or are unreachable). Since these are
not requirements for building this repository, and we don't do this
for our other tools, there is no real value in continuing to maintain
it.