Add support for 'staging' of draft service-versions. (#933)
* Add support for 'staging' of draft service versions.

This adds a 'stage' attribute to both fastly_service_vcl and
fastly_service_compute resources, with a default of 'false' (no change
to existing behavior).

If 'stage' is set to 'true', then 'plan' operations which make changes
to service versions will also 'stage' those service versions, whether
they are also activated or not. The expected usage is to have 'stage =
true' and 'activate = false', so that the provider will clone when
necessary to make draft versions, apply changes to those draft
versions and stage them, but otherwise leave them alone. When 'stage =
true' an existing draft version will *not* be cloned in order to apply
additional changes to it, they will be applied to that draft version.

Additional implementation notes:

* Added a new function in fastly_test.go to render templates (from the
  text/template module), which can be used by acceptance tests for
  generating HCL. The acceptance tests for the staging feature use
  templates.

* Moved a function from resource_fastly_service_vcl_test.go to
  fastly_test.go, refactored it, and fixed bugs in it. This function
  is used in acceptance tests to compare the list of expected backends
  for a service to the list of actual backends for the service.

* Added golang-set as a new dependency, which is used in the function
  mentioned in the last bullet.

* Apply review feedback.

* Apply suggestions from code review

Co-authored-by: Matt Cone <mattcone@users.noreply.github.com>

* Regenerate documentation.

---------

Co-authored-by: Matt Cone <mattcone@users.noreply.github.com>