Always validate VCL services after applying changes.
In PR #733 the validation logic was changed to only attempt validation
of services if they would also be activated, in order to avoid failing
validation of Compute services which do not have WASM packages
attached to them. Unfortunately this breaks customer workflows for VCL
services, as they no longer get feedback about validation failures
until they attempt to activate the service outside of Terraform.

This patch changes the logic so that VCL services are always
validated, even if they will not be activated or staged.