Upgrade to go-fastly version 11. (#1028)
* Upgrade to go-fastly version 11.

go-fastly version 11 requires a `context.Context` object to be passed
to all of its API functions; most of the provider's code already had a
Context available because the Terraform framework provides one, so
that Context was passed along. In tests where no context was already
available, `context.TODO()` was used to create one.

In addition, for resources managing services and their resources, and
for Next-Gen WAF workspaces, the go-fastly `NewContextForResourceID`
function was used to create a context unique to the specific
service/workspace being manipulated, so that go-fastly can provide
serialization of API operations for the resource.

* Add PR number to CHANGELOG.

* update go-fastly

---------

Co-authored-by: Anthony Gomez <anthony.gomez@fastly.com>