[CDTOOL-343] Add Defaults to Logging Formats in Terraform (#1066)
* init

* Linting Fix

* Escape Test

* Update Interface Test to Support Format Change for BQ

* Updated syntax to pass string

* Add Escape

* Single Test Changed for BQ

* BQ Test Changes for Review

Had to add some funky changes to support both the escape characters for TF as well as Go.

Also had to create a separate test for the Update function to ensure that we could pass in the format correctly.

* Logging Tests Refactor

Sumologic and SysLog need to be rewritten still.

* Additional Fixes and Rewrites

* OLTP Fix

* Comment corrections

* Update constants.go

* Update constants.go

* DigitalOcean and ElasticSearch Updated Defaults

Somehow missed these two. Sharing tests as well to validate.

---------

Co-authored-by: Anthony Gomez <anthony.gomez@fastly.com>