Update Datadog key validation and threshold duration defaults (#1107)
* Update Datadog key validation and threshold duration defaults

Expands the allowed length for Datadog keys from exactly 9 to between 9 and 99 characters. Makes the 'duration' field in NGWAF thresholds optional with a default of 86400 seconds, and ensures a default is set when reading if the value is zero.

* Removed Required:         false, for linter

* Duration is optional

`duration` (Number) Duration the action is in place, in seconds. Minimum 1 and maximum 31,556,900.

* Update CHANGELOG with Datadog key and NGWAF changes

Documented the expansion of allowed Datadog key lengths and the update to make the NGWAF 'duration' field optional with a default value. These changes improve flexibility and compatibility with the NGWAF API and Terraform import behavior.

* Update CHANGELOG for NGWAF fixes and enhancements

Moved details about NGWAF threshold duration and Datadog key length changes from enhancements to bug fixes, referencing PR #1107. Clarifies recent updates to the provider's behavior.

---------

Co-authored-by: Kevin P. Fleming <kpfleming@users.noreply.github.com>