Rename acceptance test functions for logging blocks (#1157)
All of the acceptance test functions for logging blocks had names
which started with the resource name, which did not cause a problem
when only one type of resource was being tested. Now that both VCL and
Compute resources are being tested, the function names have to be
changed to put that detail in the 'flavor' part of the function name.